address review: dedup ProjectVersion, drop Reader() shim, add LintReader assertion
