+++ mdl/executor/roundtrip_doctype_test.go
@@ -142,7 +142,7 @@ func TestMxCheck_DoctypeScripts(t *testing.T) {
 			}
 
 			// Filter out version-gated sections that don't match this project's Mendix version
-			pv := env.executor.Reader().ProjectVersion()
+			pv := env.executor.Backend().ProjectVersion()
 			filtered, skippedLines := filterByVersion(string(content), pv)
 			if skippedLines > 0 {
 				t.Logf("Mendix %s: skipped %d version-gated lines", pv.ProductVersion, skippedLines)
