+++ mdl/executor/roundtrip_helpers_test.go
@@ -596,7 +596,7 @@ func (e *testEnv) assertContains(createMDL string, expectedProps []string, opts 
 // requireMinVersion skips the test if the project's Mendix version is below the given minimum.
 func (e *testEnv) requireMinVersion(t *testing.T, major, minor int) {
 	t.Helper()
-	pv := e.executor.Reader().ProjectVersion()
+	pv := e.executor.Backend().ProjectVersion()
 	if !pv.IsAtLeast(major, minor) {
 		t.Skipf("Requires Mendix %d.%d+ (project is %s)", major, minor, pv.ProductVersion)
 	}
